+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.8.1
 - Instant constructors condensed: `Instant.utc`, `Instant`, `Instant.julianDate`, `Instant.dateTime`
 - Removed `from` from a lot of constructors, heavily redundant (well, maybe not so redundant with the loss of `new`)
+- Added a `timeZone` override to TimeMachine.init() so you can supply a local `DateTimeZone` to Flutter if you something
+  like `flutter_native_timezone` loaded as well.
 
 ## 0.8.0
 - Dart 2.0 ready to go. TimeMachine 0.7.1 was the last version that will work on Dart 1.24.3.